Refactored the vision code into its own class. This cleans up the code nicely.
